[ALPS08477052] change  back to do_div
[Description]
change  back  to do_div

[Test]
Mestress, ate stress

MTK-Commit-Id: dfd70a5ecaef5a7fcfeadcbe3c2a63a3ec5ee267

CR-Id: ALPS08477052
Feature: cmdq_gce

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-core.c
@@ -1598,7 +1598,7 @@ static void core_taskdone(struct work_struct *work)
 		if (perf) {
 			hw_time = perf[1] > perf[0] ?
 				perf[1] - perf[0] : ~perf[0] + 1 + perf[1];
-			CMDQ_TICK_TO_US((unsigned long long)hw_time);
+			CMDQ_TICK_TO_US(hw_time);
 		}
 	}
 	mml_mmp(exec, MMPROFILE_FLAG_END, task->job.jobid, hw_time);
